permission added.

+++ app/src/main/java/com/github/af2905/movieland/data/entities/MovieResponseDto.kt
@@ -28,6 +28,6 @@ data class MovieDto(
     @SerializedName("release_date") val releaseDate: String,
     @SerializedName("title") val title: String,
     @SerializedName("video") val video: Boolean,
-    @SerializedName("vote_average") val voteAverage: Int,
+    @SerializedName("vote_average") val voteAverage: Double,
     @SerializedName("vote_count") val voteCount: Int
 )
